ptrace: the task state is restored automatically
It's a feature of PTRACE_SEIZE.  So we need to do something, only
if we want to change the state.

[xemul: If task _was_ in stopped state before dump and we want them
 to stay alive after dump, the existing code queues one more STOP
 to it. This affects subsequent dump, as we seize a stopped task
 with STOP in queue.

 One more item in TODO list -- support stopped tasks with STOP in
 queue :)
]

+++ ptrace.c
@@ -19,13 +19,13 @@
 #include "ptrace.h"
 #include "proc_parse.h"
 
-int unseize_task(pid_t pid, int st)
+int unseize_task(pid_t pid, int orig_st, int st)
 {
 	pr_debug("\tUnseizing %d into %d\n", pid, st);
 
 	if (st == TASK_DEAD)
 		kill(pid, SIGKILL);
-	else if (st == TASK_STOPPED)
+	else if (st == TASK_STOPPED && orig_st == TASK_ALIVE)
 		kill(pid, SIGSTOP);
 	else if (st == TASK_ALIVE)
 		/* do nothing */ ;
